@@ -6,28 +6,36 @@ homepage: https://cedx.github.io/coveralls.dart
 issue_tracker: https://github.com/cedx/coveralls.dart/issues
 repository: https://github.com/cedx/coveralls.dart
 
+authors:
+  - https://cedx.github.io/coveralls.dart
+  - GiKode <827266641@qq.com>
+
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 executables:
   coveralls: coveralls
 
 dependencies:
-  build_cli_annotations: ^1.2.0
-  crypto: ^2.1.5
-  http: ^0.12.2
-  json_annotation: ^3.1.1
-  lcov: ^6.0.0
-  path: ^1.7.0
-  where: ^8.0.0
-  xml: ^4.5.1
-  yaml: ^2.2.1
+  build_cli_annotations: ^2.0.0
+  crypto: ^3.0.1
+  http: ^0.13.3
+  json_annotation: ^4.0.1
+  lcov:
+    git:
+      url: https://github.com/AugustToko/lcov.dart.git
+  path: ^1.8.0
+  where:
+    git:
+      url: https://github.com/AugustToko/where.dart.git
+  xml: ^5.1.1
+  yaml: ^3.1.0
 
 dev_dependencies:
-  build_cli: ^1.3.10
-  build_runner: ^1.11.1
-  build_test: ^1.3.6
-  build_version: ^2.0.1
-  coverage: ^0.14.2
-  json_serializable: ^3.5.1
-  test: ^1.15.7
+  build_cli: ^2.1.2
+  build_runner: ^2.0.4
+  build_test: ^2.0.4
+  build_version: ^2.0.3
+  coverage: ^1.0.3
+  json_serializable: ^4.1.3
+  test: ^1.17.5
