@@ -5,49 +5,49 @@ part of "../io.dart";
 class Job {
 
 	/// Creates a new job.
-	Job({this.repoToken, this.serviceJobId, this.serviceName, Iterable<SourceFile> sourceFiles}): sourceFiles = sourceFiles?.toList() ?? <SourceFile>[];
+	Job({this.repoToken, this.serviceJobId, this.serviceName, Iterable<SourceFile>? sourceFiles}): sourceFiles = sourceFiles?.toList() ?? <SourceFile>[];
 
 	/// Creates a new job from the specified [map] in JSON format.
 	factory Job.fromJson(Map<String, dynamic> map) => _$JobFromJson(map);
 
 	/// The current SHA of the commit being built to override the [git] property.
 	@JsonKey(name: "commit_sha")
-	String commitSha;
+	String? commitSha;
 
 	/// The job name.
 	@JsonKey(name: "flag_name")
-	String flagName;
+	String? flagName;
 
 	/// A map of Git data that can be used to display more information to users.
-	GitData git;
+	GitData? git;
 
 	/// Value indicating whether the build will not be considered done until a webhook has been sent to Coveralls.
 	@JsonKey(name: "parallel")
-	bool isParallel;
+	bool? isParallel;
 
 	/// The secret token for the repository.
 	@JsonKey(name: "repo_token")
-	String repoToken;
+	String? repoToken;
 
 	/// A timestamp of when the job ran.
 	@JsonKey(name: "run_at")
-	DateTime runAt;
+	DateTime? runAt;
 
 	/// A unique identifier of the job on the CI service.
 	@JsonKey(name: "service_job_id")
-	String serviceJobId;
+	String? serviceJobId;
 
 	/// The CI service or other environment in which the test suite was run.
 	@JsonKey(name: "service_name")
-	String serviceName;
+	String? serviceName;
 
 	/// The build number.
 	@JsonKey(name: "service_number")
-	String serviceNumber;
+	String? serviceNumber;
 
 	/// The associated pull request ID of the build.
 	@JsonKey(name: "service_pull_request")
-	String servicePullRequest;
+	String? servicePullRequest;
 
 	/// The list of source files.
 	@JsonKey(defaultValue: [], name: "source_files")
